+++ CHANGELOG.md
@@ -47,6 +47,8 @@ difference.
   compatibility.
 - Rows affected is now on by default, this is a breaking change but easily
   fixed by using the flag to turn it off.
+- Panic and Global variants are now hidden behind flags. This reduces the size
+  of the generated code by 15%. We may remove them entirely in the future.
 
 ### Removed